Fix multi-level path deps
If the top-level app has a path-based dependency (A), and that dependency also
has a path-based dependency (B) exsync would fail to start because it tried to
use the relative path from dependency B from its own working directory which
failed because the path to B is relative to A's directory.

So this commit uses `Path.expand/2` to get the correct path to the dependency.